:recycle: Rename command with cmd
