+++ cmd/root.go
@@ -1,4 +1,4 @@
-package command
+package cmd
 
 import (
 	"fmt"
