--- command/commands.go
@@ -1,1 +1,0 @@
-package command
